feat: Complete Quiz view
The functionality that allows a user to take a quiz for a deck was completed. A user can see
the question and the answer for each card of the deck and marck as correct or incorrect that
card. Then the user will be able to see the results of the quiz.